// resources/js/lib/enums/PermissionEnum.ts
// Auto-generated TypeScript enum
export enum PermissionEnum {
  CAN_ACCESS_DAHSBOARD = 'dashboard access',
  VIEW_DASHBOARD = 'view dashboard',
  VIEW_USERS = 'view users',
  CREATE_USER = 'create user',
  EDIT_USER = 'edit user',
  DELETE_USER = 'delete user',
  VIEW_USER = 'view user',
  MANAGE_USER_ROLES = 'manage user roles',
  VIEW_TEACHERS = 'view teachers',
  CREATE_TEACHER = 'create teacher',
  EDIT_TEACHER = 'edit teacher',
  DELETE_TEACHER = 'delete teacher',
  VIEW_TEACHER = 'view teacher',
  VIEW_TEACHER_WORKLOAD = 'view teacher workload',
  CHANGE_TEACHER_STATUS = 'change teacher status',
  VIEW_STUDENTS = 'view students',
  CREATE_STUDENT = 'create student',
  EDIT_STUDENT = 'edit student',
  DELETE_STUDENT = 'delete student',
  VIEW_STUDENT = 'view student',
  VIEW_INSTITUTES = 'view institutes',
  CREATE_INSTITUTE = 'create institute',
  EDIT_INSTITUTE = 'edit institute',
  DELETE_INSTITUTE = 'delete institute',
  VIEW_INSTITUTE = 'view institute',
  VIEW_DEPARTMENTS = 'view departments',
  CREATE_DEPARTMENT = 'create department',
  EDIT_DEPARTMENT = 'edit department',
  DELETE_DEPARTMENT = 'delete department',
  VIEW_DEPARTMENT = 'view department',
  VIEW_TEACHERS_WORKLOAD = 'view teachers workload',
  EXPORT_TEACHERS_WORKLOAD = 'export teachers workload',
  VIEW_TIMETABLES = 'view timetables',
  CREATE_TIMETABLE = 'create timetable',
  EDIT_TIMETABLE = 'edit timetable',
  DELETE_TIMETABLE = 'delete timetable',
  VIEW_TIMETABLE = 'view timetable',
  VIEW_ALLOCATIONS = 'view allocations',
  CREATE_ALLOCATION = 'create allocation',
  EDIT_ALLOCATION = 'edit allocation',
  DELETE_ALLOCATION = 'delete allocation',
  VIEW_ALLOCATION = 'view allocation',
  VIEW_ROOMS = 'view rooms',
  CREATE_ROOM = 'create room',
  EDIT_ROOM = 'edit room',
  DELETE_ROOM = 'delete room',
  VIEW_ROOM = 'view room',
  VIEW_SHIFTS = 'view shifts',
  CREATE_SHIFT = 'create shift',
  EDIT_SHIFT = 'edit shift',
  DELETE_SHIFT = 'delete shift',
  VIEW_SHIFT = 'view shift',
  VIEW_SLOTS = 'view slots',
  CREATE_SLOT = 'create slot',
  EDIT_SLOT = 'edit slot',
  DELETE_SLOT = 'delete slot',
  VIEW_SLOT = 'view slot',
  VIEW_PROGRAMS = 'view programs',
  CREATE_PROGRAM = 'create program',
  EDIT_PROGRAM = 'edit program',
  DELETE_PROGRAM = 'delete program',
  VIEW_PROGRAM = 'view program',
  VIEW_COURSES = 'view courses',
  CREATE_COURSE = 'create course',
  EDIT_COURSE = 'edit course',
  DELETE_COURSE = 'delete course',
  VIEW_COURSE = 'view course',
  COURSE_ATTACH_SEMESTER = 'course attach semester',
  VIEW_SEMESTERS = 'view semesters',
  CREATE_SEMESTER = 'create semester',
  EDIT_SEMESTER = 'edit semester',
  DELETE_SEMESTER = 'delete semester',
  VIEW_SEMESTER = 'view semester',
  VIEW_DAYS = 'view days',
  CREATE_DAY = 'create day',
  EDIT_DAY = 'edit day',
  DELETE_DAY = 'delete day',
  VIEW_DAY = 'view day',
  VIEW_SECTIONS = 'view sections',
  CREATE_SECTION = 'create section',
  EDIT_SECTION = 'edit section',
  DELETE_SECTION = 'delete section',
  VIEW_SECTION = 'view section',
  VIEW_ROLES = 'view roles',
  CREATE_ROLE = 'create role',
  EDIT_ROLE = 'edit role',
  DELETE_ROLE = 'delete role',
  VIEW_ROLE = 'view role',
  VIEW_PERMISSIONS = 'view permissions',
  CREATE_PERMISSION = 'create permission',
  EDIT_PERMISSION = 'edit permission',
  DELETE_PERMISSION = 'delete permission',
  VIEW_PERMISSION = 'view permission',
}